General Veers Release
Added all previously unrevealed cards from the General Veers expansion
pack, and removed the dagger marks from the cards from that pack
(because it releases this week).

Also added new Force options to allow for the construction of a Grand
Army or an "Unlimited" Force (a Force with no upper limit on any unit
type and no minimums but 1 Commander and a 7 card Command Hand).